+++ src/org/bds/test/integration/TestCasesIntegrationClusterGeneric.java
@@ -14,15 +14,15 @@ import junit.framework.Assert;
  * @author pcingola
  *
  */
-public class TestCasesClusterSsh extends TestCasesBase {
+public class TestCasesIntegrationClusterGeneric extends TestCasesBase {
 
 	@Test
-	public void test01() {
+	public void test01_log_TestCasesClusterGeneric() {
 		Gpr.debug("Test");
 
 		// Create command line
-		String[] args = { "-c", "test/clusterSsh_localhost_01.config" };
-		BdsTest bdsTest = new BdsTest("test/clusterSsh_01.bds", args, verbose, debug);
+		String[] args = { "-c", "test/clusterGeneric_localhost_01.config" };
+		BdsTest bdsTest = new BdsTest("test/clusterGeneric_01.bds", args, verbose, debug);
 		bdsTest.bds(false);
 
 		// Run script
@@ -33,9 +33,7 @@ public class TestCasesClusterSsh extends TestCasesBase {
 		// (run.pl prepends that string to PID)
 		for (Task t : bdsTest.bds.getBdsRun().getBdsThread().getTasks()) {
 			if (debug) Gpr.debug("Task " + t.getId() + ", pid " + t.getPid());
-			Assert.assertTrue("Task " + t.getId() + " was NOT executed by 'Cluster Ssh', task id " + t.getId() //
-					, t.getId().toUpperCase().startsWith("CLUSTERSSH") //
-			);
+			Assert.assertTrue("Task " + t.getId() + " was NOT executed by ClusterGeneric_localhos (pid " + t.getPid() + ")", t.getPid().startsWith("CLUSTERGENERIC_LOCALHOST_"));
 		}
 	}
 }
