+++ src/org/bds/test/integration/TestCasesIntegrationClusterSsh.java
@@ -1,4 +1,4 @@
-package org.bds.test.unit;
+package org.bds.test.integration;
 
 import org.bds.task.Task;
 import org.bds.test.BdsTest;
@@ -14,15 +14,18 @@ import junit.framework.Assert;
  * @author pcingola
  *
  */
-public class TestCasesClusterGeneric extends TestCasesBase {
+public class TestCasesIntegrationClusterSsh extends TestCasesBase {
 
+	/**
+	 * Test a simple command running via 'ssh cluster'
+	 */
 	@Test
-	public void test01_log_TestCasesClusterGeneric() {
+	public void test01() {
 		Gpr.debug("Test");
 
 		// Create command line
-		String[] args = { "-c", "test/clusterGeneric_localhost_01.config" };
-		BdsTest bdsTest = new BdsTest("test/clusterGeneric_01.bds", args, verbose, debug);
+		String[] args = { "-c", "test/clusterSsh_localhost_01.config" };
+		BdsTest bdsTest = new BdsTest("test/clusterSsh_01.bds", args, verbose, debug);
 		bdsTest.bds(false);
 
 		// Run script
@@ -33,7 +36,9 @@ public class TestCasesClusterGeneric extends TestCasesBase {
 		// (run.pl prepends that string to PID)
 		for (Task t : bdsTest.bds.getBdsRun().getBdsThread().getTasks()) {
 			if (debug) Gpr.debug("Task " + t.getId() + ", pid " + t.getPid());
-			Assert.assertTrue("Task " + t.getId() + " was NOT executed by ClusterGeneric_localhos (pid " + t.getPid() + ")", t.getPid().startsWith("CLUSTERGENERIC_LOCALHOST_"));
+			Assert.assertTrue("Task " + t.getId() + " was NOT executed by 'Cluster Ssh', task id " + t.getId() //
+					, t.getId().toUpperCase().startsWith("CLUSTERSSH") //
+			);
 		}
 	}
 }
